ExtentField.hpp: fix doxygen documentation to eliminate warning
ExtentVectorModule.hpp: fix doxygen documentation to eliminate warning

SortedIndexModule.hpp: Remove strange casts to (int) from a pointer, it's
  unnecessary (you can < and == compare pointers), it's wrong on 64 bit, and
  depending on what you think it should do, it's wrong on 32 bit since it will
  order high pointers (>2gb) below low pointers.

SortedIndexModule.cpp: Add TODO for Alistair/Craig/Brad to determine if comment
  is wrong or code is wrong.

sortedIndexTest.cpp: switch uint64 to int64 to eliminate warning; remove use of
  variable with leading _; that is against coding convention.

+++ include/DataSeries/SortedIndexModule.hpp
@@ -87,14 +87,12 @@ private:
     };
 
     static bool entrySorter(const IndexEntry *lhs, const IndexEntry *rhs) {
-        return((int)(lhs->source.get()) < (int)(rhs->source.get()) ||
-               ((int)(lhs->source.get()) == (int)(rhs->source.get()) &&
-                lhs->offset < rhs->offset));
+        return lhs->source.get() < rhs->source.get() 
+	    || (lhs->source.get() == rhs->source.get() && lhs->offset < rhs->offset);
     }
 
     static bool entryEqual(const IndexEntry *lhs, const IndexEntry *rhs) {
-        return ((int)(lhs->source.get()) == (int)(rhs->source.get()) &&
-                lhs->offset == rhs->offset);
+        return lhs->source.get() == rhs->source.get() && lhs->offset == rhs->offset;
     }
 
     typedef std::vector<IndexEntry> IndexEntryVector;
